Move classes of jflex-unicode-maven-plugin in package jflex.maven.plugin.unicode
