+++ jflex-unicode-maven-plugin/src/main/java/jflex/maven/plugin/unicode/DataFileType.java
@@ -1,4 +1,4 @@
-package jflex;
+package jflex.maven.plugin.unicode;
 
 import java.io.IOException;
 import java.io.InputStreamReader;
