+++ jflex-unicode-maven-plugin/src/main/java/jflex/maven/plugin/unicode/JFlexUnicodeMojo.java
@@ -7,7 +7,7 @@
  *                                                                         *
  * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * */
 
-package jflex;
+package jflex.maven.plugin.unicode;
 
 import java.io.File;
 import java.io.IOException;
