+++ jflex-unicode-maven-plugin/src/main/java/jflex/maven/plugin/unicode/NamedRangeSet.java
@@ -1,4 +1,4 @@
-package jflex;
+package jflex.maven.plugin.unicode;
 
 import java.util.ArrayList;
 import java.util.List;
