+++ jflex-unicode-maven-plugin/src/main/java/jflex/maven/plugin/unicode/UnicodePropertiesSkeleton.java
@@ -7,7 +7,7 @@
  *                                                                         *
  * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * */
 
-package jflex;
+package jflex.maven.plugin.unicode;
 
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
