+++ jflex-unicode-maven-plugin/src/main/jflex/ArchaicBlocksScanner.flex
@@ -1,4 +1,4 @@
-package jflex;
+package jflex.maven.plugin.unicode;
 
 import java.util.SortedSet;
 import java.util.TreeSet;
@@ -33,7 +33,7 @@ import java.util.TreeSet;
   String propertyName = "Block";
   int start;
   int end;
-  
+
   public void addPropertyValueIntervals() {
     int prevEnd = -1;
     int prevStart = -1;
@@ -57,10 +57,10 @@ import java.util.TreeSet;
 
     // Add final default property value interval, if necessary
     if (prevEnd < unicodeVersion.maximumCodePoint) {
-      unicodeVersion.addInterval(propertyName, defaultPropertyValue, 
+      unicodeVersion.addInterval(propertyName, defaultPropertyValue,
                                  prevEnd + 1, unicodeVersion.maximumCodePoint);
     }
-    
+
     // Add final named interval
     unicodeVersion.addInterval(propertyName, prevValue, prevStart, prevEnd);
   }
