+++ jflex-unicode-maven-plugin/src/main/jflex/ArchaicLineBreakScanner.flex
@@ -1,4 +1,4 @@
-package jflex;
+package jflex.maven.plugin.unicode;
 
 import java.util.SortedSet;
 import java.util.TreeSet;
