+++ jflex-unicode-maven-plugin/src/main/jflex/PropertyAliasesScanner.flex
@@ -1,4 +1,4 @@
-package jflex;
+package jflex.maven.plugin.unicode;
 
 import java.util.HashSet;
 import java.util.Set;
