+++ jflex-unicode-maven-plugin/src/main/jflex/PropertyValueAliasesScanner.flex
@@ -1,4 +1,4 @@
-package jflex;
+package jflex.maven.plugin.unicode;
 
 import java.util.HashSet;
 import java.util.Set;
