+++ jflex-unicode-maven-plugin/src/main/jflex/UnicodeDataScanner.flex
@@ -1,4 +1,4 @@
-package jflex;
+package jflex.maven.plugin.unicode;
 
 /**
  *
